Clean and split
